requirements.txt
flask
colorama
pymongo
redis
python-dotenv
